+++ packages/replay-next/components/elements-old/ElementsListData.ts
@@ -1,13 +1,13 @@
 import assert from "assert";
 import { ObjectId, PauseId, Node as ProtocolNode } from "@replayio/protocol";
 
-import { parentNodesCache } from "replay-next/components/elements/suspense/DOMParentNodesCache";
-import { Element, elementCache } from "replay-next/components/elements/suspense/ElementCache";
-import { getDistanceFromRoot } from "replay-next/components/elements/utils/getDistanceFromRoot";
-import { getItemWeight } from "replay-next/components/elements/utils/getItemWeight";
-import { isNodeInSubTree } from "replay-next/components/elements/utils/isNodeInSubTree";
-import { loadNodeSubTree } from "replay-next/components/elements/utils/loadNodeSubTree";
-import { shouldDisplayNode } from "replay-next/components/elements/utils/shouldDisplayNode";
+import { parentNodesCache } from "replay-next/components/elements-old/suspense/DOMParentNodesCache";
+import { Element, elementCache } from "replay-next/components/elements-old/suspense/ElementCache";
+import { getDistanceFromRoot } from "replay-next/components/elements-old/utils/getDistanceFromRoot";
+import { getItemWeight } from "replay-next/components/elements-old/utils/getItemWeight";
+import { isNodeInSubTree } from "replay-next/components/elements-old/utils/isNodeInSubTree";
+import { loadNodeSubTree } from "replay-next/components/elements-old/utils/loadNodeSubTree";
+import { shouldDisplayNode } from "replay-next/components/elements-old/utils/shouldDisplayNode";
 import { GenericListData } from "replay-next/components/windowing/GenericListData";
 import { recordData as recordTelemetryData } from "replay-next/src/utils/telemetry";
 import { ReplayClientInterface } from "shared/client/types";
