+++ packages/replay-next/components/elements-old/utils/getDistanceFromRoot.ts
