+++ packages/replay-next/components/elements-old/utils/shouldDisplayNode.ts
